activemq/producer/requirements.txt
stomp.py==6.0.0
CherryPy
babel
humanize
